RF: Moving stuff around a bit, and starting to cythonize voxel2fiber
