+++ dipy/tracking/_utils.py
@@ -3,6 +3,7 @@ from __future__ import division, print_function, absolute_import
 
 from warnings import warn
 import numpy as np
+from functools import wraps
 
 
 def _voxel_size_deprecated():
@@ -65,3 +66,106 @@ def _to_voxel_coordinates(streamline, lin_T, offset):
                          ' indices')
     return inds.astype(int)
 
+
+def transform_sl(sl, affine=None):
+    """
+    Helper function that moves and generates the streamline. Thin wrapper
+    around move_streamlines
+
+    Parameters
+    ----------
+    sl : list
+        A list of streamline coordinates
+
+    affine : 4 by 4 array
+        Affine mapping from fibers to data
+    """
+    if affine is None:
+        affine = np.eye(4)
+    # Generate these immediately:
+    return [s for s in move_streamlines(sl, affine)]
+
+
+def unique_rows(in_array, dtype='f4'):
+    """
+    This (quickly) finds the unique rows in an array
+
+    Parameters
+    ----------
+    in_array: ndarray
+        The array for which the unique rows should be found
+
+    dtype: str, optional
+        This determines the intermediate representation used for the
+        values. Should at least preserve the values of the input array.
+
+    Returns
+    -------
+    u_return: ndarray
+       Array with the unique rows of the original array.
+
+    """
+    x = np.array([tuple(in_array.T[:,i]) for i in
+                  xrange(in_array.shape[0])],
+        dtype=(''.join(['%s,'%dtype]* in_array.shape[-1])[:-1]))
+
+    u,i = np.unique(x, return_index=True)
+    u_i = x[np.sort(i)]
+    u_return = np.empty((in_array.shape[-1],len(u_i)))
+    for j in xrange(len(u_i)):
+        u_return[:,j] = np.array([x for x in u_i[j]])
+
+    # Return back the same dtype as you originally had:
+    return u_return.T.astype(in_array.dtype)
+
+def _with_initialize(generator):
+    """Allows one to write a generator with initialization code.
+
+    All code up to the first yield is run as soon as the generator function is
+    called and the first yield value is ignored.
+    """
+    @wraps(generator)
+    def helper(*args, **kwargs):
+        gen = generator(*args, **kwargs)
+        next(gen)
+        return gen
+
+    return helper
+
+
+@_with_initialize
+def move_streamlines(streamlines, output_space, input_space=None):
+    """Applies a linear transformation, given by affine, to streamlines.
+
+    Parameters
+    ----------
+    streamlines : sequence
+        A set of streamlines to be transformed.
+    output_space : array (4, 4)
+        An affine matrix describing the target space to which the streamlines
+        will be transformed.
+    input_space : array (4, 4), optional
+        An affine matrix describing the current space of the streamlines, if no
+        ``input_space`` is specified, it's assumed the streamlines are in the
+        reference space. The reference space is the same as the space
+        associated with the affine matrix ``np.eye(4)``.
+
+    Returns
+    -------
+    streamlines : generator
+        A sequence of transformed streamlines.
+
+    """
+    if input_space is None:
+        affine = output_space
+    else:
+        inv = np.linalg.inv(input_space)
+        affine = np.dot(output_space, inv)
+
+    lin_T = affine[:3, :3].T.copy()
+    offset = affine[:3, 3].copy()
+    yield
+    # End of initialization
+
+    for sl in streamlines:
+        yield np.dot(sl, lin_T) + offset
